backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
websockets>=12.0
opencv-python>=4.8.0
numpy>=1.24.0
PyMuPDF>=1.24.0
python-pptx>=1.0.0
Pillow>=10.0.0
onnxruntime>=1.17.0
huggingface-hub>=0.20.0
